backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
passlib[bcrypt]
email-validator
qdrant-client
requests
python-jose[cryptography]
python-multipart